requirements.txt
fastapi
uvicorn
python-socketio[asgi]
aiofiles
cohere